fix rename vertex ai
