+++ litellm/files/main.py
@@ -17,7 +17,7 @@ import litellm
 from litellm import client, get_secret_str
 from litellm.llms.azure.files.handler import AzureOpenAIFilesAPI
 from litellm.llms.openai.openai import FileDeleted, FileObject, OpenAIFilesAPI
-from litellm.llms.vertex_ai_and_google_ai_studio.files.handler import (
+from litellm.llms.vertex_ai.files.handler import (
     VertexAIFilesHandler,
 )
 from litellm.types.llms.openai import (
