+++ litellm/integrations/gcs_bucket/gcs_bucket.py
@@ -24,7 +24,7 @@ from litellm.types.utils import (
 )
 
 if TYPE_CHECKING:
-    from litellm.llms.vertex_ai_and_google_ai_studio.vertex_llm_base import VertexBase
+    from litellm.llms.vertex_ai.vertex_llm_base import VertexBase
 else:
     VertexBase = Any
