+++ litellm/llms/vertex_ai/common_utils.py
