+++ litellm/llms/vertex_ai/gemini/transformation.py
