+++ litellm/llms/vertex_ai/gemini/vertex_and_google_ai_studio_gemini.py
