+++ litellm/llms/vertex_ai/gemini_embeddings/batch_embed_content_transformation.py
