+++ litellm/llms/vertex_ai/image_generation/cost_calculator.py
