+++ litellm/llms/vertex_ai/vertex_ai_partner_models/ai21/transformation.py
