+++ litellm/llms/vertex_ai/vertex_ai_partner_models/anthropic/transformation.py
