+++ litellm/llms/vertex_ai/vertex_ai_partner_models/llama3/transformation.py
