+++ litellm/llms/vertex_ai/vertex_embeddings/transformation.py
