+++ litellm/llms/vertex_ai/vertex_embeddings/types.py
