+++ litellm/proxy/pass_through_endpoints/streaming_handler.py
@@ -13,7 +13,7 @@ from litellm.litellm_core_utils.litellm_logging import Logging as LiteLLMLogging
 from litellm.llms.anthropic.chat.handler import (
     ModelResponseIterator as AnthropicIterator,
 )
-from litellm.llms.vertex_ai_and_google_ai_studio.gemini.vertex_and_google_ai_studio_gemini import (
+from litellm.llms.vertex_ai.gemini.vertex_and_google_ai_studio_gemini import (
     ModelResponseIterator as VertexAIIterator,
 )
 from litellm.proxy._types import PassThroughEndpointLoggingResultValues
